Moved server files to cloud folder for consistency.
