+++ lib/client/workorder-client/index.js
@@ -2,7 +2,6 @@ var q = require('q');
 var _ = require('lodash');
 var shortid = require('shortid');
 var CONSTANTS = require('../../constants');
-var config = require('../../config');
 var MediatorTopicUtility = require('fh-wfm-mediator/lib/topics');
 var mediator, manager, workorderSyncSubscribers;
 
@@ -203,7 +202,7 @@ module.exports = function(_mediator) {
   mediator = _mediator;
 
   workorderSyncSubscribers = new MediatorTopicUtility(mediator);
-  workorderSyncSubscribers.prefix(CONSTANTS.SYNC_TOPIC_PREFIX).entity(config.datasetId);
+  workorderSyncSubscribers.prefix(CONSTANTS.SYNC_TOPIC_PREFIX).entity(CONSTANTS.WORKORDER_ENTITY_NAME);
 
   manager = new ManagerWrapper({
     create: create,
@@ -217,7 +216,7 @@ module.exports = function(_mediator) {
     forceSync: forceSync,
     publishRecordDeltaReceived: _.noop,
     waitForSync: waitForSync,
-    datasetId: config.datasetId
+    datasetId: CONSTANTS.WORKORDER_ENTITY_NAME
   });
 
   return manager;
