+++ lib/constants.js
@@ -2,7 +2,6 @@ module.exports = {
   TOPIC_PREFIX: "wfm",
   WORKORDER_ENTITY_NAME: "workorders",
   SYNC_TOPIC_PREFIX: "wfm:sync",
-  TOPIC_SEPARATOR: ":",
   ERROR_PREFIX: "error",
   DONE_PREFIX: "done",
   TOPICS: {
